% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datImport.R
\name{datImport}
\alias{datImport}
\title{datImport}
\usage{
datImport(RDSpath)
}
\arguments{
\item{RDSpath}{The imported dataset from google scholar using functions not
included in this package}
}
\value{
The package dataset
}
\description{
datImport
}
\examples{
RDSpath <- "inst/extdata/datIAEgeneral_newjan2022.RDS"
dat <- datImport(RDSpath)

}
