% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPhoneNumbers.R
\name{extractPhoneNumbers}
\alias{extractPhoneNumbers}
\title{Extract Phone Numbers}
\usage{
extractPhoneNumbers(inputStr)
}
\arguments{
\item{inputStr}{a character vector (supposedly) containing phone numbers
to be extracted.}
}
\value{
a list of length equal to \code{inputStr} whose entries
contain character vectors whose entries are all extracted phone
numbers from the individual entries in \code{inputStr} (if any).
}
\description{
Search for and extract all 10-digit phone numbers in a string,
provided the phone numbers are visually delimited with the typical characters
(i.e. '(', ')', '.', '+', '-', or ' ').
}
\examples{
  testStrings <- c("1234567890",
                   "123 456 7890",
                   "123.456.7890",
                   "(123) 456 7890",
                   "(123) 456 78 90",
                   "123.456.78.90",
                   "12 34 56 78 90",
                   "12.34.56.78.90",
                   "call me at 1234567890 OR 1234567890")

  extractPhoneNumbers(testStrings)
}
